Extract download function to util
